aes/pyproject.toml
[project]
name = "nalivayka"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.120.0",
    "pycryptodome>=3.23.0",
    "uvicorn>=0.38.0",
]
